# api/openapi.yaml
openapi: 3.0.0
info:
  title: Swagger Users - OpenAPI 3.0
  version: 1.0.11
...
servers:
  - url: http://localhost:3000
paths:
...
  /users/{id}:
    get:
      tags:
        - usersByID
      summary: Find user by ID
      description: Returns a single user
      operationId: getUserById
      parameters:
        - name: id
          in: path
          description: ID of user to return
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_3"
...
components:
  schemas:
...
    inline_response_200_3:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        position:
          type: string
        position_id:
          type: integer
        registration_timestamp:
          type: integer
        photo:
          type: string
      example:
        registration_timestamp: 1
        phone: phone
        name: name
        photo: photo
        id: 0
        position: position
        email: ""
        position_id: 6
...